fix(review): isolate capture lifecycle authority
